Move `run()` and `exit()` methods from `EcosystemMan` to `EcosystemController`
