+++ src/main/java/ecosim/man/EcosystemMan.java
@@ -6,10 +6,7 @@ import java.util.List;
 import java.util.logging.Level;
 
 import ecosim.Environment;
-import ecosim.SplashScreen;
 import static ecosim.common.Util.randInt;
-import static ecosim.common.io.ConsoleIO.closeConsoleInputSource;
-import static ecosim.common.io.ConsoleIO.prettyPrintln;
 import ecosim.map.Map;
 import ecosim.organism.animal.Animal;
 import ecosim.organism.animal.decorator.ConservationBoostDecorator;
@@ -46,20 +43,6 @@ public class EcosystemMan {
         // TODO: implement creating the loop for daily simulation
     }
 
-    public void run() {
-        Runtime.getRuntime().addShutdownHook(new Thread(this::exit));
-        SplashScreen.show();
-        System.out.println("Starting simulation...");
-
-        // simulator loop here
-    }
-
-    public void exit() {
-        int exitCode = 0;
-        prettyPrintln("\n<B><r>[Simulator finished w/ exit code %d]</r></B>", exitCode);
-        closeConsoleInputSource();
-    }
-
     public void createAnimal(String animal, String biome) {
         AnimalFactory animalFactory = AnimalFactoryProducer.getFactory(biome);
         Animal newAnimal = animalFactory.createAnimal(animal);
@@ -68,27 +51,28 @@ public class EcosystemMan {
         // register with the Season and TimeOfDay observables
         this.environment.registerTimeOfDayObservers(newAnimal);
         this.environment.registerSeasonObservers(newAnimal);
-        
+
         this.animals.add(decoratedAnimal);
 
     }
 
 
     private Animal decorateAnimal(Animal animal) {
-        int randomNum = randInt(0, 3); 
-        Animal decoratedAnimal = animal; 
-    
+        int randomNum = randInt(0, 3);
+        Animal decoratedAnimal = animal;
+
         switch (randomNum) {
-            case 0 -> decoratedAnimal = new ConservationBoostDecorator(animal); 
-            case 1 -> decoratedAnimal = new FertilityBoostDecorator(animal);    
-            case 2 -> decoratedAnimal = new SurvivabilityBoostDecorator(animal); 
+            case 0 -> decoratedAnimal = new ConservationBoostDecorator(animal);
+            case 1 -> decoratedAnimal = new FertilityBoostDecorator(animal);
+            case 2 -> decoratedAnimal = new SurvivabilityBoostDecorator(animal);
             default -> {
-               LoggerMan.log(Level.INFO, "Animal is returned as is (undecorated): " + animal.getName());
+                LoggerMan.log(Level.INFO, "Animal is returned as is (undecorated): " + animal.getName());
             }
         }
-    
+
         return decoratedAnimal;
     }
+
     public void createPlant(String plant, String biome) {
         PlantFactory plantFactory = PlantFactoryProducer.getFactory(biome);
         Plant newPlant = plantFactory.createPlant(plant);
@@ -107,7 +91,7 @@ public class EcosystemMan {
             map.initialisePlacement(animal);
         }
 
-        for (Plant plant : this.plants){
+        for (Plant plant : this.plants) {
             map.initialisePlacement(plant);
         }
